requirements.txt
aiohttp==3.8.3
asyncio==3.4.3
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-core==2.10.1
google-auth==2.11.1
google-auth-oauthlib==0.5.3
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-cloud-translate==3.8.2
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
numpy==1.23.3
oauthlib==3.2.1
openpyxl==3.0.10
pandas==1.5.0
plotly==5.10.0
streamlit==1.14.0
streamlit-aggrid==0.3.3
streamlit-authenticator==0.2.1
XlsxWriter==3.0.3
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
pandas-gbq==0.17.9